split CustomerServiceTests to Handlers
